move stability from tasks to flows
