--- mlip_arena/tasks/stability/__init__.py
@@ -1,1 +1,0 @@
-"""Molecular dynamics stability tasks and workflows."""
